feat: make the quiz form accessible by the finish button component and use useLocation to pass the answer to the last question to the calculating results page
