+++ src/features/quiz/QuizPage.jsx
@@ -1,10 +1,50 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useReducer } from "react";
 import ProgressBar from "./components/ProgressBar";
 import Question from "./components/Question";
 import Choices from "./components/Choices";
 import PreviousButton from "./components/PreviousButton";
 import NextButton from "./components/NextButton";
 import FinishButton from "./components/FinishButton";
+
+const initialQuizFormState = {
+  quizForm: {
+    1: null,
+    2: null,
+    3: null,
+    4: null,
+    5: null,
+    6: null,
+    7: null,
+    8: null,
+    9: null,
+    10: null,
+  },
+  selectedChoice: null,
+};
+
+function quizReducer(state, action) {
+  switch (action.type) {
+    case "SELECT_CHOICE":
+      return {
+        ...state,
+        quizForm: {
+          ...state.quizForm,
+          [action.payload.questionNumber]: action.payload.selectedLetter,
+        },
+
+        selectedChoice: action.payload.selectedLetter,
+      };
+
+    case "LOAD_SELECTED_CHOICE":
+      return {
+        ...state,
+        selectedChoice: action.payload,
+      };
+
+    default:
+      return state;
+  }
+}
 
 export default function QuizPage() {
   // total number of questions
@@ -21,6 +61,10 @@ export default function QuizPage() {
   const [enableBackButton, setEnableBackButton] = useState(false);
   const [enableFinishButton, setEnableFinishButton] = useState(false);
 
+  // this will hold all the previous choice data so it persists
+  const [state, dispatch] = useReducer(quizReducer, initialQuizFormState);
+  const { quizForm, selectedChoice } = state;
+
   useEffect(() => {
     console.log(progress);
   }, [progress]);
@@ -36,7 +80,7 @@ export default function QuizPage() {
     // check if we can enable the next button
     if (questionNumber === questionsTotal) {
       setEnableNextButton(false);
-    } 
+    }
 
     // scroll back up
     setTimeout(() => {
@@ -71,6 +115,9 @@ export default function QuizPage() {
               setEnableNextButton={setEnableNextButton}
               setEnableFinishButton={setEnableFinishButton}
               setProgress={setProgress}
+              quizForm={quizForm}
+              selectedChoice={selectedChoice}
+              dispatch={dispatch}
             />
           </div>
           <div className="flex w-full gap-5 sm:flex-row flex-col">
@@ -86,7 +133,10 @@ export default function QuizPage() {
               />
             </div>
             <div className="w-full flex justify-center sm:justify-end">
-              <FinishButton enableFinishButton={enableFinishButton} />
+              <FinishButton
+                enableFinishButton={enableFinishButton}
+                quizForm={quizForm}
+              />
             </div>
           </div>
         </div>
